libs/db-plugins-multisurvey/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "db-plugins"
version = "27.4.1a3"
description = "ALeRCE database plugins."
authors = []
readme = "README.rst"
packages = [{include = "db_plugins"}]

[tool.poetry.dependencies]
python = ">=3.7, <3.12"
alembic="1.8.1"
click="8.1.3"
SQLAlchemy="2.0.19"
pymongo="4.4.1"

[tool.poetry.group.dev.dependencies]
psycopg2-binary="*"
pytest="*"
pytest-docker="*"
pytest-cov="*"
mongomock = "^4.1.2"

[tool.poetry.group.doc.dependencies]
numpydoc="*"
recommonmark="*"

[tool.poetry.scripts]
dbp = "db_plugins.cli.manage:cli"